+++ examples/web-demo/CHANGELOG.md
@@ -1,5 +1,13 @@
 # web-demo
 
+## 0.1.42
+
+### Patch Changes
+
+- event-processor-bleeps@0.0.43
+- event-processor-conquest-eth@0.0.43
+- event-processor-nfts@0.0.43
+
 ## 0.1.41
 
 ### Patch Changes